# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "창고 PDA 시뮬레이션 API",
    "version": "1.0.0",
    "description": "WMS 창고에서 PDA(핸디터미널)가 GTL 마스터/싱글 라벨을 스캔하는 상황을 흉내낸 조회 전용 REST API. 지금은 GET만 제공하며, 이후 Flutter PDA 앱과 별도로 배포되는 다른 Vercel 프런트에서 이 API를 그대로 호출할 예정이다."
  },
  "servers": [
    {
      "url": "/",
      "description": "현재 배포"
    }
  ],
...
  "components": {
    "schemas": {
...
      "VerifyResult": {
        "type": "object",
        "description": "PDA 스캔 대조 결과 — 전량 스캔되면 출고가능",
        "properties": {
          "orderNo": {
            "type": "string",
            "example": "OUT-20260713-01"
          },
          "customer": {
            "type": "string",
            "example": "현대트랜시스"
          },
          "status": {
            "type": "string",
            "enum": [
              "출고가능",
              "미완료"
            ]
          },
          "ok": {
            "type": "boolean",
            "example": true
          },
          "expected": {
            "type": "number",
            "example": 3
          },
          "scanned": {
            "type": "number",
            "example": 3
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "아직 스캔 안 된 싱글"
          },
          "unexpected": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "주문에 없는데 스캔된 코드"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "SKU 'XXX' 재고를 찾을 수 없습니다."
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/outbound/{orderNo}/verify": {
      "post": {
        "summary": "출고 스캔 검증 (확인완료 → 출고가능)",
        "description": "PDA가 스캔한 SINGLE 라벨 목록을 해당 고객사 주문과 대조한다. 주문의 모든 싱글이 스캔되면 status가 \"출고가능\"이 된다.",
        "tags": [
          "Outbound"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderNo",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "OUT-20260713-01"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "scannedCodes"
                ],
                "properties": {
                  "scannedCodes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "GTL-SGL-0001",
                      "GTL-SGL-0006",
                      "GTL-SGL-0011"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "검증 결과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResult"
                }
              }
            }
          },
          "400": {
            "description": "요청 형식 오류 (scannedCodes 배열 필요)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "주문을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}